Restructured the code, putting the function into the utils directory to make it modular for further development
