+++ main.py
@@ -1,4 +1,4 @@
-import square
+import utils.square as square
 # This is the main application which calls the square function
 square.square()
 input("Press 'Enter' key to exit ...")
